- Marten did some general warning fixes and formating cleanup. Please people, remember the CS code guidelines. Thhere's more to be done later, be sure of that.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@24005 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csplugincommon/sndsys/snddata.h
@@ -99,7 +99,7 @@ namespace CS
     csSndSysSoundFormat m_SoundFormat;
 
     /// The number of frames in the decoded output
-    long m_FrameCount;
+    size_t m_FrameCount;
 
     /// An optional brief description of the sound data
     char *m_pDescription;
